feat(bbg_uart): improve UART communication with error handling and .gitignore
- Change `serial_write` to return `int` for better error signaling and add error handling with `perror` and `printf` for debugging.
- Add a `.gitignore` file to exclude build artifacts, OS specific files, IDE specific files, and other unnecessary files from version control.
- Update `main.c` to handle potential errors when sending data via `serial_write` and improve input handling with `scanf`.
- Add a default serial port and baud rate definition to `main.c`.

+++ include/serials.h
@@ -36,7 +36,7 @@ ssize_t serial_read(int fd, uint8_t *data, uint16_t length);
  * @param length Number of bytes to write
  * @return ssize_t number of writen bytes on success, negative -1 error code on failure
  */
-ssize_t serial_write(int fd, uint8_t *data, uint16_t length);
+int serial_write(int fd, uint8_t *data, uint16_t length);
 
 /**
  * @brief Send a packet with CRC checksum over the serial port
